content/media/video/test/test_seek5.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Media test: seek test 5</title>
  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<video id='v'
       src='seek.ogv'
       onloadedmetadata='return startTest();'
       onseeking='return seekStarted();'
       onseeked='return seekEnded();'></video>
<pre id="test">
<script class="testbody" type="text/javascript">
// Test for a seek, followed by another seek before the first is complete.
var seekCount = 0;
var completed = false;

function startTest() {
  if (completed)
    return false;

  var v = document.getElementById('v');
  v.currentTime=2;
  return false;
}

function seekStarted() {
  if (completed)
    return false;

  var v = document.getElementById('v');
  v.currentTime=1;
  return false;
}

function seekEnded() {
  if (completed)
    return false;

  var v = document.getElementById('v');
  seekCount++;
  if (seekCount == 2) {
    ok(v.currentTime >= 0.9 && v.currentTime <= 1.1, "Second seek on target: " + v.currentTime);
    completed = true;
    SimpleTest.finish();
  }

  return false;
}

SimpleTest.waitForExplicitFinish();
</script>
</pre>
</body>
</html>
